ipc-rules: fix wset-index garbage values (#2620)
* ipc-rules: fix wset-index garbage values

+++ plugins/ipc-rules/ipc-rules-common.hpp
@@ -195,8 +195,10 @@ static inline wf::json_t view_to_json(wayfire_view view)
     description["minimized"]   = toplevel ? toplevel->minimized : false;
     description["activated"]   = toplevel ? toplevel->activated : false;
     description["sticky"]     = toplevel ? toplevel->sticky : false;
-    description["wset-index"] = toplevel && toplevel->get_wset() ? toplevel->get_wset()->get_index() : -1;
-    description["min-size"]   = wf::ipc::dimensions_to_json(
+    description["wset-index"] = toplevel && toplevel->get_wset() ?
+        static_cast<int64_t>(toplevel->get_wset()->get_index()) :
+        -1;
+    description["min-size"] = wf::ipc::dimensions_to_json(
         toplevel ? toplevel->toplevel()->get_min_size() : wf::dimensions_t{0, 0});
     description["max-size"] = wf::ipc::dimensions_to_json(
         toplevel ? toplevel->toplevel()->get_max_size() : wf::dimensions_t{0, 0});
